+++ assets/AppAsset.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * @see http://www.yiiframework.com/
+ * @see       http://www.yiiframework.com/
  *
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license   http://www.yiiframework.com/license/
  */
 
 namespace app\assets;
@@ -18,7 +18,7 @@ use yii\web\YiiAsset;
  *
  * @author Qiang Xue <qiang.xue@gmail.com>
  *
- * @since 2.0
+ * @since  2.0
  */
 class AppAsset extends AssetBundle
 {
@@ -31,6 +31,7 @@ class AppAsset extends AssetBundle
     ];
 
     public $js = [
+        'js/ajax-modal-popup.js',
     ];
 
     public $depends = [