+++ console/fixtures/StaticPageFixture.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace app\console\fixtures;
+
+use app\models\StaticPage;
+use yii\test\ActiveFixture;
+
+/**
+ * Class StaticPageFixture.
+ */
+class StaticPageFixture extends ActiveFixture
+{
+    public $modelClass = StaticPage::class;
+}