+++ tests/functional/_BeforeRun.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace tests\functional;
+
+use app\console\fixtures\StaticPageFixture;
+
+abstract class _BeforeRun
+{
+    public function _fixtures()
+    {
+        return [
+            'static_page' => [
+                'class'    => StaticPageFixture::class,
+            ],
+        ];
+    }
+}